LPD-43072 separate in new exception class use same pattern as AuthorizationException
